Add proper ignoring

+++ src/main/java/cruise/umple/umpr/core/ConsoleMain.java
@@ -113,6 +113,8 @@ public class ConsoleMain {
   }
   
   private static final void removeDirectoryContents(final Path path) {
+    final Set<Path> ignorePaths = getIgnorePaths(path);
+    
     try {
       if (path.toFile().exists()) {
         Files.walkFileTree(path, new SimpleFileVisitor<Path>() {
@@ -122,13 +124,17 @@ public class ConsoleMain {
               return FileVisitResult.SKIP_SUBTREE;
             } 
             
+            if (ignorePaths.contains(dir)) {
+              return FileVisitResult.SKIP_SUBTREE;
+            }
+            
             return FileVisitResult.CONTINUE;
           }
           
           @Override
           public FileVisitResult visitFile(Path file,
               BasicFileAttributes attrs) throws IOException {
-            if (!file.toFile().isHidden()) {
+            if (!file.toFile().isHidden() && !ignorePaths.contains(file)) {
               Files.delete(file);
             }
             
@@ -158,6 +164,8 @@ public class ConsoleMain {
   
   private static void mergeDirs(final Config cfg, final Path src, final Path dest) {
 
+    final Set<Path> ignorePaths = getIgnorePaths(dest);
+    
     try {
       Files.walkFileTree(src, new SimpleFileVisitor<Path>() {
           
@@ -172,6 +180,11 @@ public class ConsoleMain {
             
             final Path relative = src.relativize(dir);
             final Path fixed = outputPath.resolve(relative);
+            
+            if (ignorePaths.contains(fixed)) {
+              return FileVisitResult.SKIP_SUBTREE;
+            }
+            
             final File fdir = fixed.toFile();
             
             fdir.mkdirs();
@@ -183,8 +196,11 @@ public class ConsoleMain {
           public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
             final Path relative = src.relativize(file);
             final Path fixed = outputPath.resolve(relative);
-            final File ffile = fixed.toFile();
+            if (ignorePaths.contains(fixed)) {
+              return FileVisitResult.SKIP_SUBTREE;
+            }
             
+            final File ffile = fixed.toFile();
             if (ffile.exists()) {
               ffile.delete();
             }
@@ -198,6 +214,32 @@ public class ConsoleMain {
     } catch (IOException ioe) {
       throw Throwables.propagate(ioe);
     }
+  }
+  
+  public static final Path IGNORE_FILE = Paths.get(".umpr.core.ignore");
+  
+  /**
+   * Gets the ignore paths from a folder, if the ignore file exists ({@value #IGNORE_FILE}). If the file exists, the 
+   * {@link Set} returned has all the paths resolved against the {@code folder} parameter including 
+   * {@value #IGNORE_FILE}.
+   * 
+   * @param folder Folder to read from
+   * @return Non-{@code null} {@link ImmutableSet} including at least {{@value #IGNORE_FILE}. 
+   */
+  private static Set<Path> getIgnorePaths(final Path folder) {
+    final Path ignPath = folder.resolve(IGNORE_FILE);
+    final File ignFile = ignPath.toFile();
+    
+    final ImmutableSet.Builder<Path> setBld = ImmutableSet.builder();
+    if (ignFile.exists()) {
+      try {
+        Files.readAllLines(ignPath).stream().map(str -> folder.resolve(Paths.get(str))).forEach(setBld::add);
+      } catch (IOException e) {
+        throw Throwables.propagate(e);
+      }
+    }
+    
+    return setBld.add(ignPath).build();
   }
  
   /**
